+++ src/features/sign-up/utils.ts
@@ -1,5 +1,6 @@
 import { SignUpInfo } from '@/types/auth';
-import * as constants from '@/pages/landing/logIn/constants';
+
+import { ERROR_MESSAGE } from '@/constants/error-message';
 
 export const checkSignUpInfo = (signUpInfo: SignUpInfo) => {
   const { id, password, passwordConfirm, nickname } = signUpInfo;
@@ -10,11 +11,11 @@ export const checkSignUpInfo = (signUpInfo: SignUpInfo) => {
     passwordConfirm.trim() === '' ||
     nickname.trim() === ''
   ) {
-    return constants.ERROR_MESSAGE.blank;
+    return ERROR_MESSAGE.blank;
   }
 
   if (password !== passwordConfirm) {
-    return constants.ERROR_MESSAGE.notSamePassword;
+    return ERROR_MESSAGE.notSamePassword;
   }
 
   if (
@@ -25,7 +26,7 @@ export const checkSignUpInfo = (signUpInfo: SignUpInfo) => {
     nickname.length < 2 ||
     nickname.length > 9
   ) {
-    return constants.ERROR_MESSAGE.invalidInput;
+    return ERROR_MESSAGE.invalidInput;
   }
 
   return '';
