+++ src/pages/logIn/LogInFail.tsx
@@ -1,8 +1,8 @@
-import * as styles from '../index.css';
+import * as styles from './index.css';
 
-import ColoredButton from '@/components/button/ColoredButton/index';
+import ColoredButton from '@/components/button/ColoredButton';
 
-import * as constants from '@/pages/landing/logIn/constants';
+import { ERROR_MESSAGE } from '@/constants/error-message';
 
 interface Props {
   closeModal: () => void;
@@ -12,9 +12,7 @@ const LogInFail = ({ closeModal }: Props) => {
   return (
     <>
       <div className={styles.logInFailMessageWrapper}>
-        <pre className={styles.logInFailMessage}>
-          {constants.ERROR_MESSAGE.loginFail}
-        </pre>
+        <pre className={styles.logInFailMessage}>{ERROR_MESSAGE.loginFail}</pre>
       </div>
       <div className={styles.buttonWrapper}>
         <ColoredButton
